change js var declare to assign for consistency across js and py translations
